pyproject.toml
[build-system]
requires=["setuptools >= 77.0.3"]
build-backend="setuptools.build_meta"

[project]
name = "NET-SEC"
version = "0.0.1"
dependencies = [
  "python-dotenv",
  "pandas",
  "numpy",
  "pymongo[srv]==4.13.1",
  "certifi",
]
requires-python = ">= 3.10"
description="NETWORK SECURITY PROJECT"
authors = [
  { name = "Ananthakrishnan K", email = "ananthakrishnan073@gmail.com" }
]
readme="README.md"

[tool.setuptools]
package-dir = {"" = "networksecurity"}

[tool.setuptools.packages.find]
where = ["networksecurity"]
